+++ src/competitions/Competition.sol
@@ -36,7 +36,7 @@ contract Competition is CompetitionInterface, DSMath, DBC, Owned {
     // Constant fields
     // Competition terms and conditions as displayed on https://ipfs.io/ipfs/QmXuUfPi6xeYfuMwpVAughm7GjGUjkbEojhNR8DJqVBBxc
     // IPFS hash encoded using http://lenschulwitz.com/base58
-    bytes public constant TERMS_AND_CONDITIONS = hex"12208E21FD34B8B2409972D30326D840C9D747438A118580D6BA8C0735ED53810491";
+    bytes public constant TERMS_AND_CONDITIONS = hex"1220B9C6706EE79792E49C633E9CB95D98E9104FB9B25D1F3D46FCC6519108251992";
     uint public MELON_BASE_UNIT = 10 ** 18;
     // Constructor fields
     address public custodian; // Address of the custodian which holds the funds sent
@@ -124,7 +124,11 @@ contract Competition is CompetitionInterface, DSMath, DBC, Owned {
     function getMelonAsset() view returns (address) { return MELON_ASSET; }
 
     /// @return Get RegistrantId from registrant address
-    function getRegistrantId(address x) view returns (uint) { return registrantToRegistrantIds[x].id; }
+    function getRegistrantId(address x) view returns (uint) {
+        bool isRegistered = registrantToRegistrantIds[x].exists;
+        require(isRegistered);
+        return registrantToRegistrantIds[x].id;
+    }
 
     /// @return Address of the fund registered by the registrant address
     function getRegistrantFund(address x) view returns (address) { return registrants[getRegistrantId(x)].fund; }