+++ tests/unit/competitions/registration.js
@@ -16,7 +16,7 @@ const BigNumber = require("bignumber.js");
 const environment = "development";
 const config = environmentConfig[environment];
 const competitionTerms =
-  "0x12208E21FD34B8B2409972D30326D840C9D747438A118580D6BA8C0735ED53810491";
+  "0xB9C6706EE79792E49C633E9CB95D98E9104FB9B25D1F3D46FCC6519108251992";
 
 const fundName = web3.utils.toHex("Super Fund");